requirements.txt
dj-database-url==0.4.1
Django==1.10
django-filter==0.14.0
djangorestframework==3.4.5
gunicorn==19.6.0
Markdown==2.6.6
psycopg2==2.6.2
virtualenv==15.0.3
whitenoise==3.2.1
